pyproject.toml
[project]
name = "tetris-clean-ppo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "tensorboard>=2.10.0",
    "wandb>=0.13.11",
    "gym==0.23.1",
    "gymnasium==0.29.1",
    "torch==2.4.1",
    "moviepy>=1.0.3",
    "pygame>=2.1",
    "huggingface-hub>=0.11.1",
    "rich<12.0",
    "tenacity>=8.2.2",
    "tyro>=0.5.10",
    "setuptools<=81.0",
    "pyyaml>=6.0.3",
    "optuna>=4.8.0",
    "optuna-dashboard>=0.20.0",
]
